Phase 2 (3.b): split ui/panes.py into a ui/panes/ package, one file per pane
Turn the single panes module into a package: base.py (RenderContext + Pane),
date.py, subway.py, hourly_weather.py, citibike.py, weather_overview.py. The
class bodies move verbatim. __init__.py re-exports the public names (with
__all__), so every existing `from ui.panes import ...` call site in layout.py,
screen.py, and the tests is unchanged. Each pane module imports `Pane` from
ui.panes.base; __init__ imports base first, so there is no circular import.

Behavior-preserving: 24 golden views at zero pixel diff; 32 tests pass.
